+++ paginas/eliminar_docentes.php
@@ -1,21 +1,26 @@
-<?php include 'navbar.php'; ?>
-<?php
-include_once '../modelo/conexion.php';
-$objeto = new Conexion();
-$conexion = $objeto->Conectar();
+<?php include 'navbar.php';
 
-if (isset($_GET['id'])) {
-    $docente_id = $_GET['id'];
+if ($rol == 1) {
+    include_once '../modelo/conexion.php';
+    $objeto = new Conexion();
+    $conexion = $objeto->Conectar();
 
-    $consulta = "DELETE FROM docentes WHERE id = :docente_id";
-    $resultado = $conexion->prepare($consulta);
-    $resultado->bindParam(':docente_id', $docente_id);
-    $resultado->execute();
+    if (isset($_GET['id'])) {
+        $docente_id = $_GET['id'];
 
-    echo "<script> alert('Docente eliminado correctamente');
+        $consulta = "DELETE FROM docentes WHERE id = :docente_id";
+        $resultado = $conexion->prepare($consulta);
+        $resultado->bindParam(':docente_id', $docente_id);
+        $resultado->execute();
+
+        echo "<script> alert('Docente eliminado correctamente');
     location.href = 'docentes.php';</script>";
+    } else {
+        echo "<script> alert('ID de docente no válido');
+location.href = 'docentes.php';</script>";
+    }
 } else {
-    echo "<script> alert('ID de docente no válido');
-    location.href = 'docentes.php';</script>";
+    echo "<script> alert('No Autorizado😡!!');
+                    location.href = 'home.php';</script>";
 }
 ?>
